requirements.txt
streamlit
Pillow
tensorflow
openai
transformers
torch 
torchvision
clip
ftfy 
regex 
tqdm